Initial refactor to use maven instead of gradle and remove all code unrelated to Graphite.  Update the dropwizard metrics library version to take advantage of bug fixes to the socket handling
