--- gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Fri Feb 27 13:45:26 CET 2015
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-2.3-bin.zip
