--- src/main/java/com/verisign/storm/metrics/util/ConnectionFailureException.java
@@ -1,25 +1,0 @@
-/*
- * Copyright 2014 VeriSign, Inc.
- *
- * VeriSign licenses this file to you under the Apache License, version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of the License at:
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is
- * distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and limitations under the License.
- *
- * See the NOTICE file distributed with this work for additional information regarding copyright ownership.
- */
-package com.verisign.storm.metrics.util;
-
-import java.io.IOException;
-
-public class ConnectionFailureException extends IOException {
-
-  public ConnectionFailureException(String s) {
-    super(s);
-  }
-
-}
